add directive
