+++ src/rest/directive.rs
@@ -1,0 +1,56 @@
+use std::collections::BTreeMap;
+
+use async_graphql::parser::types::Directive;
+use async_graphql_value::Value;
+use derive_setters::Setters;
+use serde::{Deserialize, Serialize};
+
+use crate::http::Method;
+use crate::is_default;
+
+///
+/// A structure that represents the REST directive. It allows easy parse
+#[derive(Default, Debug, Deserialize, Serialize, PartialEq, Setters)]
+pub(crate) struct Rest {
+    pub path: String,
+    #[serde(default, skip_serializing_if = "is_default")]
+    pub method: Option<Method>,
+    #[serde(default, skip_serializing_if = "is_default")]
+    pub query: BTreeMap<String, String>,
+    #[serde(default, skip_serializing_if = "is_default")]
+    pub body: Option<String>,
+}
+
+impl TryFrom<&Directive> for Rest {
+    type Error = anyhow::Error;
+
+    fn try_from(directive: &Directive) -> anyhow::Result<Self> {
+        let mut rest = Rest::default();
+
+        for (k, v) in directive.arguments.iter() {
+            if k.node.as_str() == "path" {
+                rest.path = serde_json::from_str(v.node.to_string().as_str())?;
+            }
+            if k.node.as_str() == "method" {
+                rest.method = serde_json::from_str(v.node.to_string().to_uppercase().as_str())?;
+            }
+            if k.node.as_str() == "query" {
+                if let Value::Object(map) = &v.node {
+                    for (k, v) in map {
+                        if let Value::Variable(v) = v {
+                            rest.query
+                                .insert(k.as_str().to_owned(), v.as_str().to_string());
+                        }
+                    }
+                }
+            }
+            if k.node.as_str() == "body" {
+                if let Value::Variable(v) = &v.node {
+                    rest.body = Some(v.to_string());
+                }
+            }
+        }
+
+        Ok(rest)
+    }
+}
