+++ src/rest/mod.rs
@@ -1,3 +1,4 @@
+mod directive;
 mod endpoint;
 mod endpoint_set;
 mod type_map;
