+++ src/rest/type_map.rs
@@ -5,7 +5,6 @@ use async_graphql::parser::types::Type;
 ///
 /// A intermediary data structure that allows fast access to the type of a
 /// variable by it's name.
-///
 #[derive(Debug, Clone)]
 pub struct TypeMap(BTreeMap<String, Type>);
